refactor(models): rename fused_tv_loss.py to fused_losses.py
Pure rename plus the two path references (lazy import in metrics.py, test
import). Prepares the module to host fused kernels for every dense loss,
not just TV/NLA; no behavior change.
